examples: split Hiring Pipeline (36) into multi-file project
Split monolithic main.py into two files:
- evaluators.py: specialist evaluator agents with tools and candidate data
- main.py: slim coordinator with parallel subagent mode
